Added scrolling back and forth (conveyor belt movement).
Left side press jumps left, right side press jumps right. The jumping left and right is a bit wonky.

If player goes off screen (dies), he will respawn in the middle for now.

The obstacles need to be fixed so that they spawn and despawn correctly when the screen is scrolling left and right.

Paint chips need to be fixed as well to spawn correctly.